+++ src/SimpleExporter.Sample.ConsoleApp/Program.cs
@@ -1,0 +1,32 @@
+﻿using System;
+
+namespace SimpleExporter.Sample.ConsoleApp
+{
+    internal class Program
+    {
+        private static void Main(string[] args)
+        {
+            Console.WriteLine("Sample App Started");
+            Console.WriteLine();
+
+            //Sample1
+            Console.WriteLine("-----------------");
+            Console.WriteLine("Running Sample 1");
+            Sample1.Run();
+            Console.WriteLine("Sample 1 Finish");
+            Console.WriteLine("-----------------");
+            Console.WriteLine();
+
+            //Sample2
+            Console.WriteLine("-----------------");
+            Console.WriteLine("Running Sample 2");
+            Sample2.Run();
+            Console.WriteLine("Sample 2 Finish");
+            Console.WriteLine("-----------------");
+            Console.WriteLine();
+
+            Console.WriteLine("Sample App Finish");
+            Console.ReadLine();
+        }
+    }
+}