+++ src/SimpleExporter/Writer/DelimitedTextReportWriterSetting.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SimpleExporter.Writer
+﻿namespace SimpleExporter.Writer
 {
     public class DelimitedTextReportWriterSetting : IWriterSetting
     {